Adding lightweight Memcacher class
Added a lightweight cache class for caching the responses.
So far the module always gets it's response directly from nestoria.
I added a caching class which is off by default (so this is a drop in update).
But with the right parameters it will perform caching.
So far the caching is be done with a simple hash, for bigger applications
a more sophisticated approach may be useful
(e.g. ActiveSupport::Cache::MemoryStore or FileStore)

Signed-off-by: Simon Egli <simon.egli@bbv.ch>